Users management

+++ Bytehive/Bytehive.Services/Repository/UsersRepository.cs
@@ -25,6 +25,13 @@ namespace Bytehive.Services.Repository
             this.mapper = mapper;
         }
 
+        public async Task<IEnumerable<TModel>> GetAll<TModel>()
+        {
+            return await this.db.Users
+                  .ProjectTo<TModel>(this.mapper.ConfigurationProvider)
+                  .ToListAsync();
+        }
+
         public async Task<IEnumerable<TModel>> GetAll<TModel>(string providerName)
         {
             return await this.db.Users
@@ -79,12 +86,20 @@ namespace Bytehive.Services.Repository
             return false;
         }
 
-        public async Task Remove(Guid id)
+        public async Task<bool> Delete(Guid id)
         {
             var user = await this.db.Users.FindAsync(id);
-            this.db.Users.Remove(user);
 
-            await this.db.SaveChangesAsync();
+            if (user != null)
+            {
+                this.db.Users.Remove(user);
+                await this.db.SaveChangesAsync();
+
+                return true;
+
+            }
+
+            return false;
         }
     }
 }
